# Speed up Unit.to when target unit is the same as original unit