--- a/CHANGES.rst
+++ b/CHANGES.rst
@@ -318,6 +318,9 @@ astropy.units
 - Sped up creating new composite units, and raising units to some power
   [#7549]
 
+- Sped up Unit.to when target unit is the same as the original unit.
+  [#7643]
+
 astropy.visualization
 ^^^^^^^^^^^^^^^^^^^^^