This was split out from https://github.com/astropy/astropy/pull/7616

### Before

```python
In [1]: from astropy.units import deg

In [3]: %timeit deg.to(deg)
3.69 µs ± 149 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

### After

```python
In [3]: from astropy.units import deg

In [4]: %timeit deg.to(deg)
542 ns ± 28.4 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

cc @mhvk