@mhvk 2018-07-12
@astrofrog - thanks, this PR is certainly simple and safe. But I wonder if we shouldn't go a little further. My main reason is that using `unit.to(other)` is really only useful if you want to get the scale factor, and if that is in fact what you expect, then you'll be misled if the unit conversion is not, in fact, a scaling (which would be the case with most equivalencies and with logarithmic units). I think for that case it might be better if `unit.to(other)` would just raise an exception...

Given this, I wonder if your `value=None` cannot be used for that purpose, to indicate that one just wants the scale factor if the conversion is, in fact, a simple scaling. If so, the code would be:
```
if value is None:
    return self._to(other)
else:
    <...current code...>
```
(Here, `self._to(other)` short-circuits the case of `self is other`, so this would serve your purposes.)
 
A question, then, is whether we can change the default in this way. In favour is my argument above that people are almost certainly using this to get a scale factor and thus will be getting wrong answers if equivalencies are in force. Against is that it changes the API and previously working code might now error (though, as said, I'm not sure that error isn't wanted!).

So, I think this would actually be a good change, but of course in principle we could also do something like:
```
unity = 1.
def to(self, other, value=unity, equivalencies=[]):
    if value is None:
        return self._to(other)
    if value is unity and other is self:
            return 1.
    <...> current code
```

But I must say I don't like that much. Instead, if you're unsure about my suggestion, maybe we can for now stick to the PR here, but replace `None` with `unity`, so that we ensure the default is the same yet we can make the speed-up. Indeed, another advantage of that is that it is slightly faster since the `1. if value is None else value` stanza is no longer needed. (And maybe sphinx renders is more nicely?)