@astrofrog 2018-07-12
@mhvk - I agree that for now the best solution is probably just to set the default to ``UNITY`` (which I've done in the last commit)