@mhvk 2018-07-12
See #7646 for the question of whether we should change the behaviour.